requirements.txt
opencv-python~=4.5.5.64
numpy~=1.22.3
Pillow~=9.1.0
exif~=1.3.5